AIT_Projects/requirements.txt
pandas
fastapi
spacy
numpy
transformers
torch
uvicorn
nltk
scikit-learn
black
python-dotenv